+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 * Add internal command `test-recursion`
 * Add internal command `update-formula-tests` to automatically update formula test expectations
 * Add program change log in server mode to track program updates
-* Use v2 API endpoint for reporting CPU hours
+* Use v2 API endpoints to download OEIS data and report stats
 
 ## v25.12.1